fix wazero crash on machines with low memory
Function encodeMemory used to hardcode the size of memory section to be 6.
But if TotalMemory is low, then the size of encoding of variable "pages"
is 2 bytes, not 3, so the whole module binary fails to parse:

"panic: invalid section length: expected to be 6 but got 5"
I got this error on my machine with ~962 MB of RAM (it is a VM with
dynamic memory allocation from the host). The value maxPages is 15390,
which is encoded to 9e78 (just 2 bytes).

I reworked encodeMemory to work with any amount of memory from 64 KB
(maxPages=1) to 4GB (maxPages=65536). These numbers are deduced from
the calling side (maxPages is positive, <= 65536). I added a unit test
for the function and found another edge case: maxPages < 3; for such
values of maxPages, minPages should also be updated from its hardcoded
value 3, otherwise the module fails to load, because minPages > maxPages.